+++ client/actions/modal.js
@@ -1,0 +1,6 @@
+export const showModal = ({ modalType, modalProps }) => ({
+    type: 'SHOW_MODAL', modalType, modalProps
+})
+export const hideModal = () => ({
+    type: 'HIDE_MODAL'
+})
